[organization] Move command/ -> vulpes/command/
