+++ vulpes/command/debug.hpp
